+++ src/index.js
@@ -2,17 +2,18 @@
 
 import React from "react";
 import ReactDOM from "react-dom";
-import {
-  ApolloClient,
-  createNetworkInterface,
-  ApolloProvider
-} from "react-apollo";
+import { ApolloClient } from "apollo-client";
+import { createHttpLink } from "apollo-link-http";
+import { InMemoryCache } from "apollo-cache-inmemory";
+import { ApolloProvider } from "react-apollo";
+
 import "bootstrap/dist/css/bootstrap.css";
 
 import App from "./containers/App";
 
 const client = new ApolloClient({
-  networkInterface: createNetworkInterface({ uri: "/graphql" })
+  link: createHttpLink({ uri: "/graphql" }),
+  cache: new InMemoryCache()
 });
 
 ReactDOM.render(